NVMe: Initial commit for cli program
This is a combined nvme user space program for all commands into a single
program. Usage was inspired by 'git'. Run 'nvme help' to see what commands
are available, and 'nvme help <cmd>' for detailed usage.

Some documentation is missing or partially completed, but I wanted to
get this on a public repo before leaving for holiday.

Signed-off-by: Keith Busch <keith.busch@intel.com>