NVMe: Add install target and man pages
This adds an install target, hard-coded to /usr/local/. Feel free to
propose a more correct way to do an install.

Lots of people tell me they don't have asciidoc, so they get only errors
when compiling the documenation. As such, I added the generated man pages
to the repo and made building documenation its own make target separate
from default. You can invoke "make all" or "make doc".

The only way to get to the help contents is to run "make install" now,
but it at least works correctly when you are not running nvme from the
directory you built it, so it's a step in the right direction.

Signed-off-by: Keith Busch <keith.busch@intel.com>