Added nvme-list command.
Added an new command called nvme-list which uses libudev to look for
NVM Express devices. It prints the /dev node for any detected devices
as well as some other summary information.

For now we use a Makefile hack to alter the build if libudev is not
installed. At a later date we could switch to autoconf and
./configure.