Fix English typos only
This is a commit to fix typos accross the project, mainly the
following:

  * overriden -> overridden
  * paramter -> parameter
  * Comand -> Command
  * Offical -> Official

The motivation behind this fix is to make this package perfect to be
included in the Debian project. As it is, the lintian checker is
complainging about these typos using the tag spelling-error-in-manpage.

It was also requested by a Debian maintainer in the following review:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=810033#38