fabrics: show number of controllers disconnected
Multiple controllers may have the same nqn, so we iterate multiple
times and don't want to return the status of just the last one to try to
disconnect. This patch has the status show the number of controllers that
were disconnected based on that nqn as that should be more informative.

Signed-off-by: Keith Busch <keith.busch@intel.com>