nvme: add options for write command directives
This patch includes options to specify the directive type and directive
specific. The field is meant to be used for write commands, but since
the code reuse for read, write, and compare currently exists, the options
will appear available to commands other than write.

Since the directive specific is part of the larger DSM field, this patch
also incorporates an option for the non-directive specific portion of
the DSM field.

https://github.com/linux-nvme/nvme-cli/issues/235

Signed-off-by: Keith Busch <keith.busch@gmail.com>