Allow hierarchy of similiarly named plugins
For historical reasons, we allow the user to write their command line
executing "vendor" plugin "command" as one of two ways:

  nvme vendor command

And

  nvme vendor-command

This was conflicting if a hierarchy of vendor additions exists, like
"vendor-oem".

So, this patch fixes that by first searching the entire builtin vendor
plugins before searching for matching substrings.

Signed-off-by: Keith Busch <keith.busch@intel.com>