Add a "friendly" error message on create_ns()
If you try to create a namespace when you already have reached the
maximum number of namespaces, we're throwing the following error: NVMe Status:NS_ID_UNAVAILABLE(116).
The idea of this commit is just to add a more friendly error message
together with the current message:

NVMe Status:NS_ID_UNAVAILABLE(116)
The number of namespaces supported has been exceeded!

Signed-off-by: Rodrigo R. Galvao <rosattig@linux.vnet.ibm.com>