lightnvm: fix type coercion
The lightnvm passthrough command wasn't actually aligned to the ioctl's
structure because it had an extra dword, and some compilers were having a
hard time dealing with some of the other structure casting. This fixes
the passthrough structure by removing the erroneous dword, and uses
temporary void pointers for type casting.

Link: https://github.com/linux-nvme/nvme-cli/issues/386
Signed-off-by: Keith Busch <keith.busch@intel.com>