print: Introduce show_nvme_status to print nvme status
We have so many subcommands so that we need to have a function in
common to print out status of nvme command.

The nvme status will be printed out in case the subcommand itself fails
due to a reason.  In case of command which is in the middle of the
subcommand, it will be printed as "<cmd> failed" before status.

This patch has nothing to do with functional change.

Signed-off-by: Minwoo Im <minwoo.im.dev@gmail.com>