ioctl: Do not return ioctl() directly
ioctl() will return -1 if failed so that we has to see 'errno' outside
of ioctl() call to figure out what happened in there.

As mentioned in previous commits, 'errno' sholud not be used without
system call.  This commit uses 'errno' only right after ioctl() is
returned.  Also it makes them to return the actual errno in a
negative value.

Therefore caller which is mostly nvme.c does not need to see 'errno' to
know what happened.

Signed-off-by: Minwoo Im <minwoo.im.dev@gmail.com>