property: Remove unnecessary wrapper function
nvme_get_property_helper() used to be an actual function who is getting the
value of the given property with offset.  But, Now we don't have any 'advance'
argument to get back to the caller so that we can remove the helper function by
making it nvme_get_property() which is more intuitive.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>