nvme-cli: Restore RHEL 7 compatibility
With glibc version 2.24 and before the endianness conversion macros
return a value with an incorrect type. Avoid that this causes the
nvme-cli build to fail on e.g. RHEL 7. A side effect of this patch
is that it allows sparse to verify the type of the endianness
conversion functions.

See also https://sourceware.org/bugzilla/show_bug.cgi?id=16458

Cc: Mikhail Skorzhinskii <mskorzhinskiy@solarflare.com>
Cc: Minwoo Im <minwoo.im.dev@gmail.com>
Reported-by: Mikhail Skorzhinskii <mskorzhinskiy@solarflare.com>
Fixes: e68409e447b2 ("Remove superfluous casts")
Signed-off-by: Bart Van Assche <bvanassche@acm.org>