Use a systemd app-specific machine ID for hostnqn
If /etc/nvme/hostnqn is not present, the fabric commands will ask
systemd for an app-specific machine ID as a fallback.  This should
improve functionality if /etc/nvme/hostnqn is not present and should
allow packagers to avoid creating /etc/nvme/hostnqn.

Heavily based on an earlier patch from Tomasz Torcz.

Signed-off-by: Tomasz Torcz <tomek@pipebreaker.pl>
Signed-off-by: Andy Lutomirski <luto@kernel.org>