Initial libnvme commit
Most of the planned features are developed, however many of the interfaces have
only been lighlty touched as the features they implement are not broadly
supported in hardware.

The remaining gaps in command support from 1.4 are:

  Management Interface Send/Recieve OOB support
  Persistent Event Logs

There is no plan to support Abort, AER, IO SQ/CQ Create/Delete, or Doorbell
Buffer Config since that has high risk to abuse and interfere with the
stability of a running system.

Signed-off-by: Keith Busch <kbusch@kernel.org>