Import ccan modules
Replace <sys/queue.h> list implementation with the superior ccan list.
This is a more pleasant C implementation to work with, and since we the
structures using this are not exported through the API, we can freely
change these kinds of implementation details.

And while we're at it, include a few other useful modules to provide
useful macros and compile time checks on spec defined structs.

Signed-off-by: Keith Busch <kbusch@kernel.org>