Add man pages for exports
Generate man-pages from existing source. This will likely change soon
as the types (struct and enum) should probably be grouped in the same
manual as the primary function that uses it rather than have their own,
and similiar functions can in turn be grouped in the same manual as
well.

Signed-off-by: Keith Busch <kbusch@kernel.org>