nvme-print: fix offset calculation in NS ID Descriptor list
The offset for the next entry in the list is based on the length (len)
of the previous entry, but len is erroneously accumulated instead of
being set. This works fine for the first and second entry in the list
but will fail to reach the correct offset for the 3rd entry and beyond.

Signed-off-by: Saar Gross <saarg@amazon.com>