nvme-print: notify hexadecimal format to nsid header
The following two different commands print different format of Namespace
ID.

	nvme list:		Prints decimal nsid
	nvme list --verbose:	Prints hexadecimal nsid without 0x prefix.

It might cause confusions to users for figuring out namespace IDs.
Also, it might not possible to update the print-format because there
might many systems that parse the output of this command as a
hexadecimal.  If we put the prefix '0x' to the value, it might cause
some problems for systems to parse that value based on the previous
format.

This patch describes the format of the output value to the header:

	NSID(hex)
	------------
	48

It makes clear that the output value is in hexadecimal format.

Signed-off-by: Minwoo Im <minwoo.im.dev@gmail.com>