{get,set}_feature: get nsid from device node
If the given device is the controller device, set nsid to 0xFFFFFFFF.
This allows something like

  nvme set-feature /dev/nvme0 -f 0x5 -v $((1 << 16))

to work as expected.

Since we cannot use S_ISCHR/S_ISBLK to differentiate between controller
and namespace devices since commit e77046661509 ("remove block device
checking for namespace"), just try the NVME_IOCTL_ID ioctl and if it
fails with ENOTTY, assume that it is the controller device.

Signed-off-by: Klaus Jensen <k.jensen@samsung.com>