id-ctrl vs for Amazon Elastic Block Store
This allows the user to extract useful information from the
vendor-specific field when querying Elastic Block Storage
devices attached by Amazon Web Services. In particular, this
field is populated with the block device that was requested
by the user in the API call, which is of the form `/dev/sdc`
or `/dev/xvdc` rather than `/dev/nvme2`. This plugin reduces
the need for operating system maintainers to include vendor-
specific packages in their distribution.

https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/nvme-ebs-volumes.html
https://github.com/aws/amazon-ec2-utils/blob/master/ebsnvme-id