nvme-print: fix ZNS MAR/MOR print format
MAR is currently printed as unsigned.
MOR is currently printed as signed.
They should use the same print format.

Printing MOR as unsigned on a controller that has no limit,
shows as -1, which doesn't make sense.

Since no limit is represented as 0xffffffff in the ZNS spec,
printing MAR/MOR as hex makes the most sense.

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>